requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
firebase-admin==5.2.0
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
python-jose==3.3.0
python-multipart==0.0.5
opencv-python==4.5.5.64
torch==2.2.0
Pillow==10.4.0
requests==2.32.3
